moved files concerned with ising calculation to ising folder
